error blocks
there's been a few errors in the code. i've commented them out and
labelled them off with a ~Josh bookmark. if you ctrl+F that you can find
exactly where they are. i'm currently looking into finding out why it's
not working like that. im thinking it's the '='. maybe implement a set
method to get rid of this error?